+++ App/Zotero/extensions/zoteroOpenOfficeIntegration@zotero.org/components/zoteroOpenOfficeIntegration.js
@@ -404,7 +404,8 @@ Application.prototype = {
 	secondaryFieldType: "Bookmark",
 	supportedNotes: ["footnote", "endnote"],
 	supportsImportExport: true,
-	outputFormat: "rtf"
+	outputFormat: "rtf",
+	processorName: "LibreOffice"
 };
 
 /**