+++ App/Zotero/extensions/zoteroWinWordIntegration@zotero.org/components/zoteroWinWordIntegration.js
@@ -53,20 +53,20 @@ function init() {
 	document_t = new ctypes.StructType("document_t");
 	
 	field_t = new ctypes.StructType("field_t", [
-		{ "code":ctypes.jschar.ptr },
-		{ "text":ctypes.jschar.ptr },
-		{ "noteType":ctypes.unsigned_short },
-		{ "bookmarkName":ctypes.jschar.ptr },
-		{ "textLocation":ctypes.long },
-		{ "noteLocation":ctypes.long }
+		{ code: ctypes.jschar.ptr },
+		{ text: ctypes.jschar.ptr },
+		{ noteType: ctypes.unsigned_short },
+		{ bookmarkName: ctypes.jschar.ptr },
+		{ textLocation: ctypes.long },
+		{ noteLocation: ctypes.long }
 		// There's more here, but we will never access it, and we do not create field_t objects
 		// from JavaScript
 	]);
 	
 	fieldListNode_t = new ctypes.StructType("fieldListNode_t");
 	fieldListNode_t.define([
-		{ "field":field_t.ptr },
-		{ "next":fieldListNode_t.ptr }
+		{ field: field_t.ptr },
+		{ next: fieldListNode_t.ptr }
 	]);
 	
 	progressFunction_t = new ctypes.FunctionType(ctypes.stdcall_abi, ctypes.void_t,
@@ -75,96 +75,104 @@ function init() {
 	var statusCode = ctypes.unsigned_short;
 	f = {
 		// void clearError(void);
-		"clearError":lib.declare("clearError", ctypes.stdcall_abi, ctypes.void_t),
+		clearError: lib.declare("clearError", ctypes.stdcall_abi, ctypes.void_t),
 		
 		// jschar* getError(void);
-		"getError":lib.declare("getError", ctypes.stdcall_abi, ctypes.jschar.ptr),
+		getError: lib.declare("getError", ctypes.stdcall_abi, ctypes.jschar.ptr),
 		
 		// statusCode getDocument(const jschar* documentName, Document** returnValue);
-		"getDocument":lib.declare("getDocument", ctypes.stdcall_abi, statusCode,
+		getDocument: lib.declare("getDocument", ctypes.stdcall_abi, statusCode,
 			ctypes.jschar.ptr, document_t.ptr.ptr),
 		
 		// void freeDocument(Document *doc);
-		"freeDocument":lib.declare("freeDocument", ctypes.stdcall_abi, statusCode, document_t.ptr),
+		freeDocument: lib.declare("freeDocument", ctypes.stdcall_abi, statusCode, document_t.ptr),
 		
 		// statusCode displayAlert(jschar const dialogText[], unsigned short icon,
 		//						   unsigned short buttons, unsigned short* returnValue);
-		"displayAlert":lib.declare("displayAlert", ctypes.stdcall_abi, ctypes.unsigned_short,
+		displayAlert: lib.declare("displayAlert", ctypes.stdcall_abi, ctypes.unsigned_short,
 			document_t.ptr, ctypes.jschar.ptr, ctypes.unsigned_short, ctypes.unsigned_short,
 			ctypes.unsigned_short.ptr),
 		
 		// statusCode canInsertField(Document *doc, const jschar fieldType[], bool* returnValue);
-		"canInsertField":lib.declare("canInsertField", ctypes.stdcall_abi, statusCode,
+		canInsertField: lib.declare("canInsertField", ctypes.stdcall_abi, statusCode,
 			document_t.ptr, ctypes.jschar.ptr, ctypes.bool.ptr),
 		
 		// statusCode cursorInField(Document *doc, const jschar fieldType[], Field** returnValue);
-		"cursorInField":lib.declare("cursorInField", ctypes.stdcall_abi, statusCode, document_t.ptr,
+		cursorInField: lib.declare("cursorInField", ctypes.stdcall_abi, statusCode, document_t.ptr,
 			ctypes.jschar.ptr, field_t.ptr.ptr),
 		
 		// statusCode getDocumentData(Document *doc, jschar **returnValue);
-		"getDocumentData":lib.declare("getDocumentData", ctypes.stdcall_abi, statusCode,
+		getDocumentData: lib.declare("getDocumentData", ctypes.stdcall_abi, statusCode,
 			document_t.ptr, ctypes.jschar.ptr.ptr),
 		
 		// statusCode setDocumentData(Document *doc, const jschar documentData[]);
-		"setDocumentData":lib.declare("setDocumentData", ctypes.stdcall_abi, statusCode,
+		setDocumentData: lib.declare("setDocumentData", ctypes.stdcall_abi, statusCode,
 			document_t.ptr, ctypes.jschar.ptr),
 		
 		// statusCode insertField(Document *doc, const jschar fieldType[],
 		//					      unsigned short noteType, Field **returnValue)
-		"insertField":lib.declare("insertField", ctypes.stdcall_abi, statusCode, document_t.ptr,
+		insertField: lib.declare("insertField", ctypes.stdcall_abi, statusCode, document_t.ptr,
 			ctypes.jschar.ptr, ctypes.unsigned_short, field_t.ptr.ptr),
 		
 		// statusCode getFields(document_t *doc, const jschar fieldType[],
 		//					    fieldListNode_t** returnNode);
-		"getFields":lib.declare("getFields", ctypes.stdcall_abi, statusCode, document_t.ptr,
+		getFields: lib.declare("getFields", ctypes.stdcall_abi, statusCode, document_t.ptr,
 			ctypes.jschar.ptr, fieldListNode_t.ptr.ptr),
 		
 		// statusCode setBibliographyStyle(Document *doc, long firstLineIndent, 
 		//								   long bodyIndent, unsigned long lineSpacing,
 		//								   unsigned long entrySpacing, long tabStops[],
 		//								   unsigned long tabStopCount);
-		"setBibliographyStyle":lib.declare("setBibliographyStyle", ctypes.stdcall_abi,
+		setBibliographyStyle: lib.declare("setBibliographyStyle", ctypes.stdcall_abi,
 			statusCode, document_t.ptr, ctypes.long, ctypes.long, ctypes.unsigned_long,
 			ctypes.unsigned_long, ctypes.long.array(), ctypes.unsigned_long),
+
+		// statusCode exportDocument(Document *doc, const jschar fieldType[], const jschar importInstructions[]);
+		exportDocument: lib.declare("exportDocument", ctypes.stdcall_abi, statusCode, document_t.ptr,
+			ctypes.jschar.ptr, ctypes.jschar.ptr),
+
+		// statusCode importDocument(Document *doc, const jschar fieldType[], bool *returnValue);
+		importDocument: lib.declare("importDocument", ctypes.stdcall_abi, statusCode, document_t.ptr,
+			ctypes.jschar.ptr, ctypes.bool.ptr),
 		
 		// statusCode convert(document_t *doc, field_t* fields[], unsigned long nFields,
 		//				      const jschar toFieldType[], unsigned short noteType[]);
-		"convert":lib.declare("convert", ctypes.stdcall_abi, statusCode, document_t.ptr,
+		convert: lib.declare("convert", ctypes.stdcall_abi, statusCode, document_t.ptr,
 			field_t.ptr.ptr, ctypes.unsigned_long, ctypes.jschar.ptr, ctypes.unsigned_short.ptr),
 		
 		// statusCode cleanup(Document *doc);
-		"cleanup":lib.declare("cleanup", ctypes.stdcall_abi, statusCode, document_t.ptr),
+		cleanup: lib.declare("cleanup", ctypes.stdcall_abi, statusCode, document_t.ptr),
 
 		// statusCode complete(Document *doc);
-		// "complete":lib.declare("complete", ctypes.stdcall_abi, statusCode, document_t.ptr),
+		// complete:lib.declare("complete", ctypes.stdcall_abi, statusCode, document_t.ptr),
 		
 		// statusCode deleteField(Field* field);
-		"deleteField":lib.declare("deleteField", ctypes.stdcall_abi, statusCode, field_t.ptr),
+		deleteField: lib.declare("deleteField", ctypes.stdcall_abi, statusCode, field_t.ptr),
 			
 		// statusCode removeCode(Field* field);
-		"removeCode":lib.declare("removeCode", ctypes.stdcall_abi, statusCode, field_t.ptr),
+		removeCode: lib.declare("removeCode", ctypes.stdcall_abi, statusCode, field_t.ptr),
 			
 		// statusCode selectField(Field* field);
-		"selectField":lib.declare("selectField", ctypes.stdcall_abi, statusCode, field_t.ptr),
+		selectField: lib.declare("selectField", ctypes.stdcall_abi, statusCode, field_t.ptr),
 			
 		// statusCode setText(Field* field, const jschar string[], bool isRich);
-		"setText":lib.declare("setText", ctypes.stdcall_abi, statusCode, field_t.ptr,
+		setText: lib.declare("setText", ctypes.stdcall_abi, statusCode, field_t.ptr,
 			ctypes.jschar.ptr, ctypes.bool),
 			
 		// statusCode getText(Field* field, jschar** returnValue);
-		"getText":lib.declare("getText", ctypes.stdcall_abi, statusCode, field_t.ptr,
+		getText: lib.declare("getText", ctypes.stdcall_abi, statusCode, field_t.ptr,
 			ctypes.jschar.ptr.ptr),
 			
 		// statusCode setCode(Field *field, const jschar code[]);
-		"setCode":lib.declare("setCode", ctypes.stdcall_abi, statusCode, field_t.ptr,
+		setCode: lib.declare("setCode", ctypes.stdcall_abi, statusCode, field_t.ptr,
 			ctypes.jschar.ptr),
 		
 		// statusCode getNoteIndex(Field* field, unsigned long *returnValue);
-		"getNoteIndex":lib.declare("getNoteIndex", ctypes.stdcall_abi, statusCode,
+		getNoteIndex: lib.declare("getNoteIndex", ctypes.stdcall_abi, statusCode,
 			field_t.ptr, ctypes.unsigned_long.ptr),
 		
 		// statusCode freeData(void* ptr);
-		"freeData":lib.declare("freeData", ctypes.stdcall_abi, statusCode, ctypes.void_t.ptr)
+		freeData: lib.declare("freeData", ctypes.stdcall_abi, statusCode, ctypes.void_t.ptr)
 	};
 	
 	fieldPtr = new ctypes.PointerType(field_t);
@@ -213,8 +221,8 @@ Application.prototype = {
 	classID:		Components.ID("{c7a7eec5-f073-4db4-b383-f866f4b96b1c}"),
 	contractID:		"@zotero.org/Zotero/integration/application?agent=WinWord;1",
 	QueryInterface: XPCOMUtils.generateQI([Components.interfaces.nsISupports]),
-	"service":		true,
-	"getDocument":function(documentName) {
+	service:		true,
+	getDocument: async function(documentName) {
 		init();
 		var docPtr = new document_t.ptr();
 		// For OneDrive stored documents we receive the in the form:
@@ -226,11 +234,15 @@ Application.prototype = {
 		checkStatus(f.getDocument(documentName, docPtr.address()));
 		return new Document(docPtr);
 	},
-	"getActiveDocument":function(path) {
+	getActiveDocument: async function(path) {
 		return this.getDocument(null);
 	},
-	"primaryFieldType":"Field",
-	"secondaryFieldType":"Bookmark"
+	primaryFieldType: "Field",
+	secondaryFieldType: "Bookmark",
+	supportedNotes: ["footnote", "endnote"],
+	supportsImportExport: true,
+	outputFormat: "rtf",
+	processorName: "Word"
 };
 
 /**
@@ -241,7 +253,7 @@ var Document = function(cDoc) {
 	this._documentStatus = {"active":true};
 };
 Document.prototype = {
-	"displayAlert":function(dialogText, icon, buttons) {
+	displayAlert: function(dialogText, icon, buttons) {
 		Zotero.debug("ZoteroWinWordIntegration: displayAlert", 4);
 		var buttonPressed = new ctypes.unsigned_short();
 		checkStatus(f.displayAlert(this._document_t, dialogText, icon, buttons,
@@ -249,9 +261,9 @@ Document.prototype = {
 		return buttonPressed.value;
 	},
 	
-	"activate":function() {},
+	activate: function() {},
 	
-	"canInsertField":function(fieldType) {
+	canInsertField: function(fieldType) {
 		Zotero.debug("ZoteroWinWordIntegration: canInsertField", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new ctypes.bool();
@@ -259,7 +271,7 @@ Document.prototype = {
 		return returnValue.value;
 	},
 	
-	"cursorInField":function(fieldType) {
+	cursorInField: function(fieldType) {
 		Zotero.debug("ZoteroWinWordIntegration: cursorInField", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new field_t.ptr();
@@ -267,7 +279,7 @@ Document.prototype = {
 		return (returnValue.isNull() ? null : new Field(returnValue, this._documentStatus));
 	},
 	
-	"getDocumentData":function() {
+	getDocumentData: function() {
 		Zotero.debug("ZoteroWinWordIntegration: getDocumentData", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new ctypes.jschar.ptr();
@@ -277,13 +289,13 @@ Document.prototype = {
 		return data;
 	},
 	
-	"setDocumentData":function(documentData) {
+	setDocumentData: function(documentData) {
 		Zotero.debug(`ZoteroWinWordIntegration: setDocumentData ${documentData}`, 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.setDocumentData(this._document_t, documentData));
 	},
 	
-	"insertField":function(fieldType, noteType) {
+	insertField: function(fieldType, noteType) {
 		Zotero.debug("ZoteroWinWordIntegration: insertField", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new field_t.ptr();
@@ -291,40 +303,52 @@ Document.prototype = {
 		return new Field(returnValue, this._documentStatus);
 	},
 	
-	"getFields":function(fieldType) {
+	getFields: async function(fieldType) {
 		Zotero.debug("ZoteroWinWordIntegration: getFields", 4);
 		checkIfFreed(this._documentStatus);
 		var fieldListNode = new fieldListNode_t.ptr();
 		checkStatus(f.getFields(this._document_t, fieldType, fieldListNode.address()));
-		return new FieldEnumerator(fieldListNode, this._documentStatus);
+		var fnum = new FieldEnumerator(fieldListNode, this._documentStatus);
+		var fields = [];
+		while (fnum.hasMoreElements()) {
+			fields.push(fnum.getNext());
+			await Zotero.Promise.delay();
+		}
+		return fields;
 	},
 	
-	"getFieldsAsync":function(fieldType, observer) {
-		Zotero.debug("ZoteroWinWordIntegration: getFieldsAsync", 4);
-		observer.observe(this.getFields(fieldType), "fields-available", null);
-	},
-	
-	"setBibliographyStyle":function(firstLineIndent, bodyIndent, lineSpacing, entrySpacing,
+	setBibliographyStyle: function(firstLineIndent, bodyIndent, lineSpacing, entrySpacing,
 			tabStops) {
 		Zotero.debug("ZoteroWinWordIntegration: setBibliographyStyle", 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.setBibliographyStyle(this._document_t, firstLineIndent, bodyIndent, lineSpacing,
 			entrySpacing, ctypes.long.array(tabStops.length)(tabStops), tabStops.length));
 	},
+
+	importDocument: function(fieldType) {
+		Zotero.debug(`ZoteroWinWordIntegration: importDocument`, 4);
+		checkIfFreed(this._documentStatus);
+		var returnValue = new ctypes.bool();
+		checkStatus(f.importDocument(this._document_t, fieldType, returnValue.address()));
+		return returnValue.value;
+	},
+
+	exportDocument: function(fieldType, importInstructions) {
+		Zotero.debug(`ZoteroWinWordIntegration: exportDocument`, 4);
+		checkIfFreed(this._documentStatus);
+		checkStatus(f.exportDocument(this._document_t, fieldType, importInstructions));
+	},
 	
-	"convert":function(fieldEnumerator, toFieldType, toNoteTypes, nFields) {
+	convert: function(fields, toFieldType, toNoteTypes, nFields) {
 		Zotero.debug("ZoteroWinWordIntegration: convert", 4);
 		checkIfFreed(this._documentStatus);
-		var fieldPointers = [];
-		while(fieldEnumerator.hasMoreElements()) {
-			fieldPointers.push(fieldEnumerator.getNext()._field_t);
-		}
-		checkStatus(f.convert(this._document_t, field_t.ptr.array()(fieldPointers),
-			fieldPointers.length, ctypes.jschar.array()(toFieldType),
+		fields = fields.map(field => field._field_t);
+		checkStatus(f.convert(this._document_t, field_t.ptr.array()(fields),
+			fields.length, ctypes.jschar.array()(toFieldType),
 			ctypes.unsigned_short.array()(toNoteTypes)));
 	},
 	
-	"cleanup":function() {
+	cleanup: function() {
 		Zotero.debug("ZoteroWinWordIntegration: cleanup", 4);
 		if(this._documentStatus.active) {
 			checkStatus(f.cleanup(this._document_t));
@@ -333,7 +357,7 @@ Document.prototype = {
 		}
 	},
 	
-	"complete":function() {
+	complete: function() {
 		Zotero.debug("ZoteroWinWordIntegration: complete", 4);
 		if(this._documentStatus.active) {
 			// checkStatus(f.complete(this._document_t));
@@ -353,12 +377,12 @@ var FieldEnumerator = function(startNode, documentStatus) {
 	this._documentStatus = documentStatus;
 };
 FieldEnumerator.prototype = {
-	"hasMoreElements":function() {
+	hasMoreElements: function() {
 		checkIfFreed(this._documentStatus);
 		return !this._currentNode.isNull();
 	},
 	
-	"getNext":function() {
+	getNext: function() {
 		checkIfFreed(this._documentStatus);
 		var contents = this._currentNode.contents;
 		var fieldPtr = contents.addressOfField("field").contents;
@@ -379,31 +403,31 @@ var Field = function(field_t, documentStatus) {
 	this._documentStatus = documentStatus;
 };
 Field.prototype = {
-	"delete":function() {
+	delete: function() {
 		Zotero.debug("ZoteroWinWordIntegration: delete", 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.deleteField(this._field_t));
 	},
 	
-	"removeCode":function() {
+	removeCode: function() {
 		Zotero.debug("ZoteroWinWordIntegration: removeCode", 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.removeCode(this._field_t));
 	},
 	
-	"select":function() {
+	select: function() {
 		Zotero.debug("ZoteroWinWordIntegration: select", 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.selectField(this._field_t));
 	},
 	
-	"setText":function(text, isRich) {
+	setText: function(text, isRich) {
 		Zotero.debug("ZoteroWinWordIntegration: setText", 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.setText(this._field_t, text, isRich));
 	},
 	
-	"getText":function(text) {
+	getText: function() {
 		Zotero.debug("ZoteroWinWordIntegration: getText", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new ctypes.jschar.ptr();
@@ -411,13 +435,13 @@ Field.prototype = {
 		return returnValue.readString();
 	},
 	
-	"setCode":function(code) {
+	setCode: function(code) {
 		Zotero.debug("ZoteroWinWordIntegration: setCode "+code, 4);
 		checkIfFreed(this._documentStatus);
 		checkStatus(f.setCode(this._field_t, code));
 	},
 	
-	"getCode":function(code) {
+	getCode: function() {
 		Zotero.debug("ZoteroWinWordIntegration: getCode", 4);
 		checkIfFreed(this._documentStatus);
 		var code = this._field_t.contents.addressOfField("code").contents.readString();
@@ -425,7 +449,7 @@ Field.prototype = {
 		return code;
 	},
 	
-	"equals":function(field) {
+	equals: function(field) {
 		Zotero.debug("ZoteroWinWordIntegration: equals", 4);
 		checkIfFreed(this._documentStatus);
 		// Obviously, a field cannot be equal to a bookmark
@@ -445,12 +469,23 @@ Field.prototype = {
 		}
 	},
 	
-	"getNoteIndex":function(field) {
+	getNoteIndex: function() {
 		Zotero.debug("ZoteroWinWordIntegration: getNoteIndex", 4);
 		checkIfFreed(this._documentStatus);
 		var returnValue = new ctypes.unsigned_long();
 		checkStatus(f.getNoteIndex(this._field_t, returnValue.address()));
 		return parseInt(returnValue.value);
+	}
+}
+
+for (let cls of [Document, Field]) {
+	for (let method in cls.prototype) {
+		if (typeof cls.prototype[method] == 'function') {
+			let syncMethod = cls.prototype[method];
+			cls.prototype[method] = async function() {
+				return syncMethod.apply(this, arguments);
+			}
+		}
 	}
 }